@@ -25,7 +25,7 @@ var (
 	class   string
 	queries []string
 	objects []string
-	options rest.Options
+	withRules bool
 
 	limit                 int
 	since, until, timeout time.Duration
@@ -35,7 +35,7 @@ func startFlags(cmd *cobra.Command) {
 	cmd.Flags().StringArrayVarP(&queries, "query", "q", nil, "Query string for start objects, can be multiple.")
 	cmd.Flags().StringVar(&class, "class", "", "Class for serialized start objects")
 	cmd.Flags().StringArrayVar(&objects, "object", nil, "Serialized start object, can be multiple.")
-	cmd.Flags().BoolVar(&options.Rules, "rules", false, "Include rule names in returned graph")
+	cmd.Flags().BoolVar(&withRules, "rules", false, "Include rule names in returned graph")
 }
 
 func constraintFlags(cmd *cobra.Command) {
@@ -85,7 +85,7 @@ var (
 			defer cancel()
 			g, err := traverse.New(e, e.Graph()).Neighbours(ctx, start(e), depth)
 			check(err)
-			newPrinter(os.Stdout).Print(rest.NewGraph(g, options))
+			newPrinter(os.Stdout).Print(rest.NewGraph(g, withRules))
 		},
 	}
 	depth int
@@ -113,7 +113,7 @@ var (
 			defer cancel()
 			g, err := traverse.New(e, e.Graph()).Goals(ctx, start(e), goals)
 			check(err)
-			newPrinter(os.Stdout).Print(rest.NewGraph(g, options))
+			newPrinter(os.Stdout).Print(rest.NewGraph(g, withRules))
 		},
 	}
 )
