@@ -12,10 +12,10 @@ import (
 
 // Constraint included in a store Get operation to restrict the resulting objects.
 type Constraint struct {
-	Limit   *int           `json:"limit,omitempty"`                                                                 // Limit number of objects returned per query, <=0 means no limit.
-	Timeout *time.Duration `json:"timeout,omitempty" swaggertype:"string"`                                          // Timeout per request, h/m/s/ms/ns format
-	Start   *time.Time     `json:"start,omitempty" swaggertype:"string" format:"date-time" extensions:"x-nullable"` // Start of time interval, quoted RFC 3339 format.
-	End     *time.Time     `json:"end,omitempty" swaggertype:"string" format:"date-time" extensions:"x-nullable"`   // End of time interval, quoted RFC 3339 format.
+	Limit   *int           `json:"limit,omitempty"`   // Limit number of objects returned per query.
+	Timeout *time.Duration `json:"timeout,omitempty"` // Timeout per request, h/m/s/ms/ns format
+	Start   *time.Time     `json:"start,omitempty"`   // Start of time interval (RFC 3339).
+	End     *time.Time     `json:"end,omitempty"`     // End of time interval (RFC 3339).
 }
 
 func (c *Constraint) String() string {
